Add ErrKeepAndRetryTaskLater
This will cause a retry but, when possible, it will avoid requeueing
the message and instead allow the same message to be processed again.

Also fix a bug so that retriable errors don't get recorded as an error
(it appears this was the intended behavior, and it's definitely what
I want).