Improve SQS broker reliability when backend returns errors
Originally, the SQS broker created a goroutine with an infinite loop. That
goroutine would poll SQS and put any received messages on a channel. A separate
goroutine would poll that channel, spawning a goroutine for any received items.
A separate channel reported errors between these different goroutines. In
theory, all the goroutines would listen for channel closure and error channel
items to handle errors. Unfortunately, if the backend reports errors then
polling the queue would quickly stop, because in this case the currency pool
wasn't properly returned.

To address this, the multiple goroutines are collapsed. When the initial
pollying goroutine receives a message, it immediately spawns a goroutine to
process that message, rather than using a channel and intermediate goroutine.
In addition, it always returns to the concurrency pool immediately after either
encountering an error or processing a message. This allows polling and
processing to continue even if the backend reports an error.